perf(tinygrad): switch pipeline wire format from fp32 to bf16
bf16 preserves fp32's range (no saturation on large residual stream
outlier features, unlike fp16) while using half the bandwidth of
fp32. For a 3072-hidden model shipping one hidden state per decode
token that's ~3 KB/step instead of ~6 KB/step.

Numpy doesn't expose bf16 natively, so the wire carries uint16 bit
patterns and tinygrad does the actual dtype conversion:

  sender:   t.cast(bfloat16).bitcast(uint16).numpy() -> uint16 ndarray
  wire:     <shape header><raw bf16 bytes>
  receiver: Tensor(uint16_arr).bitcast(bfloat16).cast(activation_dtype)

encode_hidden / decode_hidden switch dtype assertions from fp32 to
uint16. send_hidden / recv_hidden signatures updated accordingly.
Tests use np.uint16 arrays for the wire fixtures (semantically
arbitrary 16-bit patterns since the transport is byte-transparent);
test fixtures for model weights stay fp32.

+++ src/exo/worker/engines/tinygrad/tests/test_pipeline_group.py
@@ -30,7 +30,7 @@ def test_encode_decode_token_roundtrip() -> None:
 
 
 def test_encode_decode_hidden_roundtrip() -> None:
-    arr = np.arange(12, dtype=np.float32).reshape(1, 3, 4)
+    arr = np.arange(12, dtype=np.uint16).reshape(1, 3, 4)
     buf = encode_hidden(arr)
     out = decode_hidden(buf)
     assert out.shape == arr.shape  # pyright: ignore[reportAny]
@@ -83,7 +83,7 @@ def test_two_rank_ring_localhost_token_roundtrip() -> None:
     assert token == 99 and not stop
 
     # Hidden state: rank 0 -> rank 1
-    arr = np.arange(24, dtype=np.float32).reshape(1, 6, 4)
+    arr = np.arange(24, dtype=np.uint16).reshape(1, 6, 4)
     groups[0].send_hidden(arr)
     recv = groups[1].recv_hidden()
     assert recv.shape == arr.shape  # pyright: ignore[reportAny]
